Include unit tests for book.rs resources inside the module
